Added auto hide/show Setting & Help like Material Design guidelines say:
http://www.google.com/design/spec/patterns/navigation-drawer.html (Check last point)

If your drawer content is larger, the method that I have implemented will show the sticky buttons, if not, it auto hides it.

You can try this deleting or adding items inside drawer.xml (now I have a lot of "the drawer content" elements, remove some of them to get the result with a scrollView that can't scroll).